Link against -lstdc++ to fix an error with STL collection hooks.
I tried to avoid this but this was the best way, and should be fine until I can
come up with a better solution. It's risky because we may pull in other stuff
but I'm hoping I can control that, for now.

Might be nice to just find the actual implementations?